requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pytest
pytest-asyncio